requirements.txt
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
gql==3.5.0
python-jose==3.3.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
plotly==5.18.0
pandas==2.1.4
gunicorn==21.2.0
